+++ docs/changelog.rst
@@ -38,6 +38,7 @@ fixes
 - Delete variable to reset the forecast date for test
 - Fixed and adjust the internal Colombia shape for maps
 - Fixed internal thresholds for PPT in 3 categories
+- Fix number format in climatology table
 
 
 0.6.0 (**2013-06-19**)